refactor(camera_config): update warmup duration type and adjust Xense camera configuration
- Changed the type of `warmup_s` in both OpenCVCameraConfig and RealSenseCameraConfig from `int` to `float` for improved precision in warmup duration.
- Added a new "front" camera configuration to the XenseMultisensorConfig, enhancing camera setup options.
- Removed the redundant "front" camera configuration to streamline the code.

These changes improve camera configuration flexibility and maintainability across the project.

+++ src/lerobot/cameras/opencv/configuration_opencv.py
@@ -60,7 +60,7 @@ class OpenCVCameraConfig(CameraConfig):
     index_or_path: int | Path
     color_mode: ColorMode = ColorMode.RGB
     rotation: Cv2Rotation = Cv2Rotation.NO_ROTATION
-    warmup_s: int = 1
+    warmup_s: float = 1.0
     fourcc: str | None = "MJPG"
 
     def __post_init__(self) -> None:
